Make caasp4os clean more idempotent
Don't fail with make clean in the unlikely event of a CTRL-Ced run of make
caasp4os, where we didn't populate a buildir/deployment yet.

+++ backend/caasp4os/destroy.sh
@@ -27,10 +27,12 @@ if [ -d "$BUILD_DIR" ]; then
         wait
     fi
 
-    pushd deployment
-    # destroy terraform openstack stack
-    skuba_container terraform destroy -auto-approve
-    popd
+    if [ -d deployment ]; then
+        pushd deployment
+        # destroy terraform openstack stack
+        skuba_container terraform destroy -auto-approve
+        popd
+    fi
 
     popd
     rm -rf "$BUILD_DIR"
